+++ src/main/resources/data.sql
@@ -1,5 +1,6 @@
-/*INSERT into users (username, password, email, role, dateAccount) values ('user', 'user1234', 'user@mail.com', 'USER');
-INSERT into users (username, password, email, role) values ('admin', 'admin1234', 'admin@mail.com', 'ADMIN');*/
+INSERT into users (username, password, email, role, date_account) values ("user", "user1234", "user@mail.com", "USER", '2021-08-10 08:43:34.645954');
+INSERT into users (username, password, email, role, date_account) values ('admin', 'admin1234', 'admin@mail.com', 'ADMIN', '2021-08-10 08:43:34.645954');
+
 
 INSERT into post (title, content, username, created_on) values ("Le logiciel espion Pegasus et les faiblesses des Européens", "S’ils s’en étaient pris à la Chine, à la Russie ou aux Etats-Unis, le risque de représailles aurait été massif pour les utilisateurs du logiciel espion Pegasus, de la société israélienne NSO Group. En revanche, « s’ils visaient l’Europe, ils savaient que rien ne se passerait ». Le diagnostic tranchant de Frédéric Mauro, avocat et spécialiste de la défense, coauteur de Défendre l’Europe (Nuvis, 2019), est partagé par plusieurs spécialistes à Bruxelles, après les révélations du Monde et de seize médias concernant la surveillance ciblée de dirigeants, d’activistes et de journalistes européens. Ursula von der Leyen, présidente de la Commission européenne, a réagi avec une déclaration d’usage, qualifiant les faits « d’inacceptables, s’ils sont avérés ». Au Conseil (l’institution représentant les Etats), on se partage entre fatalisme et inquiétude. Une « cellule de prospective » a été récemment chargée d’établir la carte des nouvelles menaces pesant sur l’Union. Et de ses indéniables faiblesses. « Le cyber est dans la situation du contre-terrorisme européen il y a dix ans. Or cette lutte est encore plus difficile et implique encore plus d’acteurs, alors qu’on n’identifie pas clairement qui est le pilote des opérations », relève une source au Conseil.
